Fix: library overrides moved from mediaView to mediaOverrides
